+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+1.0.0
+* Add save & remove functions
+
+0.1.3
+* Fix mobile display bug.
+
+0.1.2
+* Enable hotspot interactivity.
+* Enable multiple hotspots.
+* Fix known bugs.
+
 0.1.1
 * Fix usage in README.